third_party/blink/web_tests/platform/generic/fast/dom/SelectorAPI/caseTagX-expected.txt

PASS document.querySelector('div SPAN') is null
PASS document.getElementById('lower1').webkitMatchesSelector('div SPAN') is false
PASS successfullyParsed is true

TEST COMPLETE
